Extracted inner classes out of ActionLineParsers and moved utility methods to TokenisedLine
