+++ tichu-cli/src/main/java/net/incongru/tichu/simu/parse/ActionLineParser.java
@@ -1,0 +1,9 @@
+package net.incongru.tichu.simu.parse;
+
+import net.incongru.tichu.action.Action;
+
+interface ActionLineParser {
+    boolean accept(TokenisedLine t);
+
+    Action parse(TokenisedLine t);
+}
